added float support in summ (examples_002)

+++ examples/examples_002/Program.cs
@@ -1,7 +1,7 @@
 ﻿Console.Write("Введите первое число  ");
-int a = Int32.Parse(Console.ReadLine());
+float a = float.Parse(Console.ReadLine());
 Console.Write("Введите второе число   ");
-int b = Int32.Parse(Console.ReadLine());
-int summ = a + b;
+float b = float.Parse(Console.ReadLine());
+float summ = a + b;
 Console.Write("Сумма этих чисел равна   ");
 Console.WriteLine(summ);
